lib: rename PluginRegistry
As we have so many kinds of plugins, we'll just call it
ImplementationFactory as it is a factory that returns implementations.
