+++ core/src/stored/dev.cc
@@ -122,14 +122,15 @@ Device* FactoryCreateDevice(JobControlRecord* jcr,
   Dmsg1(400, "max_block_size in device_resource res is %u\n",
         device_resource->max_block_size);
 
-  if (!PluginRegistry<Device>::IsRegistered(device_resource->dev_type)) {
+  if (!ImplementationFactory<Device>::IsRegistered(device_resource->dev_type)) {
     Jmsg2(jcr, M_ERROR, 0, _("%s has an unknown device type %s\n"),
           device_resource->archive_device_string,
           device_resource->dev_type.c_str());
     return nullptr;
   }
 
-  Device* dev = PluginRegistry<Device>::Create(device_resource->dev_type);
+  Device* dev
+      = ImplementationFactory<Device>::Create(device_resource->dev_type);
 
   dev->device_resource = device_resource;
   device_resource->dev = dev;
